refactor: decouple  CoreState to rari_core
