+++ crates/rari/src/server/middleware/mod.rs
@@ -1,3 +1,3 @@
 pub mod proxy;
-pub mod request;
+pub use rari_core::middleware::request;
 pub use rari_core::request_context;
