+++ crates/rari_core/Cargo.toml
@@ -94,6 +94,8 @@ deno_kv = { workspace = true, optional = true }
 deno_cron = { workspace = true, optional = true }
 deno_node_sqlite = { workspace = true, optional = true }
 hyper-util = { version = "0.1.20", features = ["client-legacy", "tokio"] }
+tokio-tungstenite = "0.30.0"
+tungstenite = "0.30.0"
 libc = "0.2.186"
 nix = { workspace = true, features = ["term"] }
 node_resolver = { workspace = true }
@@ -101,6 +103,9 @@ redb = "4.1.0"
 redis = { version = "1.3.0", default-features = false, features = ["tokio-comp"] }
 rustyline = { workspace = true }
 sys_traits = { workspace = true }
+
+[dev-dependencies]
+async-stream = { version = "0.3.6" }
 
 [lints]
 workspace = true
