+++ crates/rari_core/src/cache/mod.rs
@@ -1,4 +1,5 @@
 //! Byte-level cache handler abstraction shared across backends.
 pub mod handler;
+pub mod response;
 
 pub use handler::*;
