+++ crates/rari_core/src/lib.rs
@@ -20,3 +20,7 @@ pub mod runtime;
 pub mod sanitize;
 pub mod component;
 pub mod action_state;
+pub mod rendering;
+pub mod middleware;
+pub mod vite;
+pub mod state;
