+++ crates/rari_core/src/middleware/mod.rs
@@ -1,0 +1,2 @@
+//! HTTP middleware shared across backends.
+pub mod request;
