+++ crates/rari_core/src/rendering/mod.rs
@@ -1,0 +1,4 @@
+//! Generic HTML/response rendering utilities shared across backends.
+pub mod pretty_html;
+pub mod streaming_response;
+pub mod utils;
