+++ crates/rari_core/src/rendering/streaming_response.rs
